module5-task1 - Перенеет логику отрисовки фильмов в контролер
